Rename db/request to db/requests
